% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pal.R
\name{d_pal}
\alias{d_pal}
\alias{data_pal}
\title{Colours for data values}
\usage{
d_pal(x, pal = hcl.colors(84))

data_pal(x, pal = hcl.colors(84))
}
\arguments{
\item{x}{data vector, maybe be numeric or character}

\item{pal}{palette, may be colours or a function}
}
\description{
Scales input data to the palette, so that colour is mapped linearly to the range of values.
}
\details{
Default palette 'pal' is the 'viridis' colours of [grDevices::hcl.colors()], and may be literal colour values
or a function.

[data_pal()] is an alias of [d_pal()].
}
\examples{
plot(1:100, col = d_pal(1:100))
plot(1:100, col = d_pal(1:100, chl_pal))

}
