% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty.R
\name{dirty_image}
\alias{dirty_image}
\title{Convert list of raster output to bathymetry palette}
\usage{
dirty_image(x, alpha = 1)
}
\arguments{
\item{x}{object returned by vapour package 'gdal_raster_data()'}

\item{alpha}{0 to 1 for transparency}
}
\value{
modified x
}
\description{
Convert list of raster output to bathymetry palette
}
