% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty.R
\name{dirty_pal}
\alias{dirty_pal}
\title{Bathymetry palette, absolute for global use}
\usage{
dirty_pal(x, palette = FALSE, alpha = 1)
}
\arguments{
\item{x}{a vector of data values or a single num (n)}

\item{palette}{logical, if \code{TRUE} return a list with matching colours and values}

\item{alpha}{0 to 1 for transparency}
}
\value{
colours, palette, or function, see Details
}
\description{
From AAD underway facility (DiRT).
}
\details{
The palette functions operate in 3 modes:
1) n colours - Pal(6) - returns 6 colours from the palette
2) data      - Pal(c(-4000, -2000, 100)) - return colours for 3 elevations
3) palette   - Pal(palette = TRUE) - return the full palette and breaks

The palette spans -8000 to 1000 metres, values outside that range clamp to
the end colours.
}
\examples{
dirty_pal(seq(-8000, 0, length.out = 10))
}
