% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datapal.R
\name{ice.pal}
\alias{ice.pal}
\title{Sea ice colours}
\usage{
ice.pal(x, palette = FALSE, alpha = 1)
}
\arguments{
\item{x}{a vector of data values or a single num (n)}

\item{palette}{logical, if \code{TRUE} return a list with matching colours and values}

\item{alpha}{value in 0,1 to specify opacity}
}
\value{
colours, palette, or function, see Details
}
\description{
Colours for sea ice.
}
\details{
The palette functions operate in 3 modes:
1) n colours - ice.pal(6) - returns 6 colours from the palette
2) data      - ice.pal(c(10, 50, 100)) - return colours for 3 ice concentrations
3) palette   - ice.pal(palette = TRUE) - return the full palette and breaks
}
\references{
Derived from \url{http://www.iup.uni-bremen.de/seaice/amsr/README_GEOTIFF.txt}.
}

