% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePal.R
\name{mk_timePal}
\alias{mk_timePal}
\title{Create a time-index colour map.}
\usage{
mk_timePal(x, col)
}
\arguments{
\item{x}{date-times}

\item{col}{colours, can be a function or an actual set of colours}
}
\value{
function of date-time
}
\description{
Create a time-index colour map.
}

