% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palr-package.r
\docType{package}
\name{palr}
\alias{palr-package}
\alias{palr}
\title{palr}
\description{
palr: colours for data
}
\seealso{
Useful links:
\itemize{
  \item \url{https://australianantarcticdivision.github.io/palr/}
  \item \url{https://github.com/AustralianAntarcticDivision/palr}
  \item Report bugs at \url{https://github.com/AustralianAntarcticDivision/palr/issues}
}

}
\author{
\strong{Maintainer}: Michael D. Sumner \email{mdsumner@gmail.com}

Other contributors:
\itemize{
  \item Abigael Proctor (Named the package) [contributor]
  \item Tomas Remenyi (Provided colours for element_pal) [contributor]
  \item R Core Team and contributors worldwide (source code of image.default) [contributor]
}

}
