% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palr.R
\name{sst_pal}
\alias{sst_pal}
\alias{sstPal}
\title{SST colours}
\usage{
sst_pal(x, palette = FALSE, alpha = 1, ...)

sstPal(x, palette = FALSE, alpha = 1, ...)
}
\arguments{
\item{x}{a vector of data values or a single number}

\item{palette}{logical, if \code{TRUE} return a list with matching colours and values}

\item{alpha}{value in 0,1 to specify opacity}

\item{...}{currently ignored}
}
\value{
colours, palette, or function, see Details
}
\description{
SST colours
}
\examples{
library(raster)
data(oisst)
sstcols <- sst_pal(palette = TRUE)
plot(oisst, col = sstcols$col, zlim = range(sstcols$breaks))
}
\references{
Derived from a file once found at 'http://oceancolor.gsfc.nasa.gov/DOCS/palette_sst.txt'
}
