% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dirty.R
\name{dirty_pal}
\alias{dirty_pal}
\title{Bathymetry palette, absolute for global use}
\usage{
dirty_pal(x, alpha = 1)
}
\arguments{
\item{x}{numeric value}

\item{alpha}{0 to 1 for transparency}
}
\value{
vector of hex colours
}
\description{
From AAD underway facility.
}
\examples{
dirty_pal(seq(-8000, 0, length.out = 10))
}
