% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/timePal.R
\name{mk_timePal}
\alias{mk_timePal}
\title{Create a time-index colour map.}
\usage{
mk_timePal(x, col)
}
\arguments{
\item{x}{date-times}

\item{col}{colours, can be a function or an actual set of colours}
}
\value{
function of date-time
}
\description{
Create a time-index colour map.
}
\examples{
library(trip)
library(adehabitatLT)
tr <- as.data.frame(as.trip(albatross))
trs <- split(tr, tr$id)
palfun <- mk_timePal(tr$date, colorRampPalette(c("red", "blue")))
plot(tr$x, tr$y, pch = ".")
for (i in seq_along(trs)) {
  for (j in seq(nrow(trs[[i]]))[-1]) {
    lines(trs[[i]][(j-1):j, c("x", "y")], col = palfun(trs[[i]]$date[j]))
  }
}
}

