% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/palr.R
\docType{data}
\name{oisst}
\alias{oisst}
\title{Sample NOAA 1/4 degree daily Optimum Interpolation Sea Surface Temperature}
\format{\code{oisst} A RasterLayer object}
\description{
This is a gridded raster data set of SST.
}
\references{
\url{http://www.ncdc.noaa.gov/oisst}
}
\keyword{data}

